requirements.txt
Flask
requests
lxml
requests-pkcs12
zeep
python-dotenv
pytesseract
Pillow
playwright
p12importer